2023/5/22 23:17

+++ app/src/main/java/com/example/currentweatherforecast/bean/weatherbean/WeatherHourInfo.java
@@ -76,7 +76,7 @@ public class WeatherHourInfo implements Serializable {
         @SerializedName("dt")
         public Integer dt;
         @SerializedName("precipitation")
-        public Integer precipitation;
+        public Double precipitation;
     }
 
     @NoArgsConstructor
